requirements.txt
google_spreadsheet
google-auth-oauthlib
matplotlib
astropy
